refactor(consensus/adaptor): remove dead code
Remove two entirely-dead subsystems and several unreachable functions
kept alive only by their own coverage tests:

- Delete txpool.go (TxPool struct + methods): the live abstraction is
  the consensus.TxPool interface; this concrete type had no caller.
- Delete unl.go's dead UNL type, NewUNL, CalcQuorum, DecodeValidatorKey;
  preserve the sole live function (DecodeValidatorKeyWithMaster) by
  moving it beside its caller ParseValidatorKeysWithMaster in startup.go.
- Delete the dead ParseValidatorKeys wrapper (production calls
  ParseValidatorKeysWithMaster directly).

Relocate test-only helpers out of production source into a new
testhelpers_test.go so they no longer ship in the binary:
TransactionToMessage, HaveSetToMessage, fetchPackCache.size,
newLedgerProviderForTest, NewRouterBroadcaster, and the txSet reply-cap
test setters.

Drop redundant ParseValidatorKeys wrapper tests (the WithMaster variants
already cover the same paths); preserve the focused key-decode tests
retargeted to DecodeValidatorKeyWithMaster.

Verified: full module builds, adaptor suite green, deadcode reports no
production-unreachable functions in the package.
