+++ internal/consensus/adaptor/converter.go
@@ -99,15 +99,6 @@ func TransactionFromMessage(msg *message.Transaction) []byte {
 	return msg.RawTransaction
 }
 
-// TransactionToMessage wraps a raw transaction blob into a Transaction message.
-func TransactionToMessage(txBlob []byte) *message.Transaction {
-	return &message.Transaction{
-		RawTransaction:   txBlob,
-		Status:           message.TxStatusNew,
-		ReceiveTimestamp: uint64(time.Now().UnixNano()),
-	}
-}
-
 // HaveSetFromMessage converts a decoded HaveTransactionSet message.
 func HaveSetFromMessage(msg *message.HaveTransactionSet) (consensus.TxSetID, message.TxSetStatus) {
 	var id consensus.TxSetID
@@ -115,14 +106,6 @@ func HaveSetFromMessage(msg *message.HaveTransactionSet) (consensus.TxSetID, mes
 		copy(id[:], msg.Hash)
 	}
 	return id, msg.Status
-}
-
-// HaveSetToMessage creates a HaveTransactionSet message.
-func HaveSetToMessage(id consensus.TxSetID, status message.TxSetStatus) *message.HaveTransactionSet {
-	return &message.HaveTransactionSet{
-		Status: status,
-		Hash:   id[:],
-	}
 }
 
 func xrplEpochToTime(epoch uint32) time.Time {
